program3/requirements.txt
aws-cdk-lib==2.181.0
constructs~=10.4.2
mypy-extension==1.0.0
pathspec==0.12.1
platformdirs==4.3.6